Split integration test
